pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for Crawl4AI with RAG capabilities and knowledge graph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.7.0",
    "supabase>=2.0.0",
    "openai>=1.0.0",
    "fastmcp>=2.0.0",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "sentence-transformers>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0,<14.0",
    "psutil>=5.9.0",
    "playwright>=1.55.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "knowledge_graphs"]
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # Line length handled by Black
]
extend-exclude = [".venv", "build", "dist", "__pycache__"]

[tool.ruff.per-file-ignores]
"tests/*.py" = ["F401", "F811"]  # Allow unused imports in tests
